# openapi.json
{"openapi":"3.1.0","info":{"title":"API de Análise de Contratos","summary":"Trabalho final da disciplina de Construção de APIs para IA","description":"API que disponibiliza endpoints de Inteligência Artificial para extração e análise de informações em documentos contratuais.","termsOfService":"http://example.com/terms/","contact":{"name":"José Allan - Elson - PH - Geiziane","url":"https://github.com/joseallangoncalves/fastapi_contratos","email":"joseallan@ufg.br"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0.html"},"version":"0.1"},"paths":{"/api/v1/health":{"get":{"tags":["Health Check"],"summary":"Perform a Health Check","description":"## Perform a Health Check\nEndpoint to perform a healthcheck on. This endpoint can primarily be used Docker\nto ensure a robust container orchestration and management is in place. Other\nservices which rely on proper functioning of the API service will not deploy if this\nendpoint returns any other HTTP status code except 200 (OK).\nReturns:\n    HealthCheck: Returns a JSON response with the health status","operationId":"get_health_api_v1_health_get","responses":{"200":{"description":"Return HTTP Status Code 200 (OK)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthCheck"}}}}}}},"/api/v1/token":{"post":{"tags":["Autenticação"],"summary":"Login For Access Token","operationId":"login_for_access_token_api_v1_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_api_v1_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/contratos/analise-contratual":{"post":{"tags":["Análise de Contratos (IA)"],"summary":"Análise de Multas e Penalidades Contratuais","description":"Recebe um contrato em PDF, extrai cláusulas específicas de penalidades, multas, sanções e prazos de vigência utilizando IA e salva os dados de forma estruturada no banco de dados.","operationId":"analise_contratual_api_v1_contratos_analise_contratual_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_analise_contratual_api_v1_contratos_analise_contratual_post"}}},"required":true},"responses":{"201":{"description":"Análise realizada e salva no banco de dados com sucesso.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnaliseContratualResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/contratos/analise-sms":{"post":{"tags":["Análise de Contratos (IA)"],"summary":"Análise de Requisitos SMS","description":"Recebe um anexo/contrato em PDF e extrai obrigações de Segurança, Meio Ambiente e Saúde aplicáveis à Contratada.","operationId":"analise_sms_api_v1_contratos_analise_sms_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_analise_sms_api_v1_contratos_analise_sms_post"}}},"required":true},"responses":{"201":{"description":"Obrigações de SMS extraídas com sucesso.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnaliseSMSResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/contratos/gerar-checklist":{"post":{"tags":["Análise de Contratos (IA)"],"summary":"Geração de Checklist de Fiscalização","description":"Recebe um contrato em PDF e gera automaticamente um checklist de verificação para apoiar o processo de recebimento e fiscalização.","operationId":"gerar_checklist_api_v1_contratos_gerar_checklist_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_gerar_checklist_api_v1_contratos_gerar_checklist_post"}}},"required":true},"responses":{"201":{"description":"Checklist gerado com sucesso.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GerarChecklistResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/contratos/analises":{"get":{"tags":["Análise de Contratos (IA)"],"summary":"Listar Análises Contratuais Persistidas","description":"Retorna uma lista de todas as análises de contratos salvas no banco de dados.\nPermite filtragem opcional por empresa ou CNPJ.","operationId":"listar_analises_api_v1_contratos_analises_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"empresa","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Empresa"}},{"name":"cnpj","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cnpj"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/contratos/analises/{id}":{"get":{"tags":["Análise de Contratos (IA)"],"summary":"Obter Detalhes de uma Análise Específica","description":"Retorna os detalhes completos (incluindo todas as cláusulas salvas) de uma análise contratual específica por ID.","operationId":"obter_analise_api_v1_contratos_analises__id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AnaliseContratualDadosSalvos":{"properties":{"empresa":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Empresa","description":"Nome da empresa contratada extraído do documento"},"cnpj":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cnpj","description":"CNPJ da empresa contratada extraído do documento"},"data_inicio":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Data Inicio","description":"Data de início de vigência do contrato"},"data_fim":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Data Fim","description":"Data de fim de vigência do contrato"},"valor_contrato":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Valor Contrato","description":"Valor global do contrato"},"data_insercao":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Data Insercao","description":"Data e hora em que o registro foi persistido no banco"}},"type":"object","title":"AnaliseContratualDadosSalvos"},"AnaliseContratualResponse":{"properties":{"id_analise":{"type":"integer","title":"Id Analise","description":"Identificador único da análise salva no banco de dados"},"analise":{"type":"string","title":"Analise","description":"Relatório de análise de multas e penalidades gerado pela IA em Markdown"},"dados_salvos":{"$ref":"#/components/schemas/AnaliseContratualDadosSalvos","description":"Metadados estruturados extraídos e salvos"}},"type":"object","required":["id_analise","analise","dados_salvos"],"title":"AnaliseContratualResponse"},"AnaliseSMSResponse":{"properties":{"analise":{"type":"string","title":"Analise","description":"Relatório de obrigações de Saúde, Segurança e Meio Ambiente (SMS) gerado em Markdown"}},"type":"object","required":["analise"],"title":"AnaliseSMSResponse"},"Body_analise_contratual_api_v1_contratos_analise_contratual_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_analise_contratual_api_v1_contratos_analise_contratual_post"},"Body_analise_sms_api_v1_contratos_analise_sms_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_analise_sms_api_v1_contratos_analise_sms_post"},"Body_gerar_checklist_api_v1_contratos_gerar_checklist_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_gerar_checklist_api_v1_contratos_gerar_checklist_post"},"Body_login_for_access_token_api_v1_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_api_v1_token_post"},"GerarChecklistResponse":{"properties":{"analise":{"type":"string","title":"Analise","description":"Checklist estruturado para fiscalização e recebimento contratual em Markdown"}},"type":"object","required":["analise"],"title":"GerarChecklistResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthCheck":{"properties":{"status":{"type":"string","title":"Status","default":"OK"}},"type":"object","title":"HealthCheck","description":"Response model to validate and return when performing a health check."},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"api/v1/token"}}}}}}